+++ R/input_indicator_create.R
@@ -14,7 +14,7 @@
 #'                              year = 1990, unit = "MIO_EUR", 
 #'                              households = FALSE, labelling = "iotables")
 #'
-#'de_emp_indicator <- input_indicator_create ( de_emp, de_output)
+#'de_emp_indicator <- input_indicator_create ( de_emp, de_output )
 #' @export
 
 input_indicator_create <- function ( input_matrix, output_vector,
@@ -46,8 +46,9 @@ input_indicator_create <- function ( input_matrix, output_vector,
     }
    
   }
+  
   input_matrix[,1] <- as.character (input_matrix[,1])
   input_matrix[1,1] <- paste0(as.character(input_matrix[1,1]), "_indicator")
   
-  return(input_matrix)
+  input_matrix
 }